Interchange foreign and frozen state codes
Users upgrading from a previous version of slackroll to the new one
will be better if their old custom packages are marked as frozen
instead of foreign, so frozen must have the same code as the old
custom state.

If custom->frozen, customized version of official packages as well
as other packages like aaa_elflibs will be frozen after the upgrade,
which is desirable because it will prevent them from being upgraded
automatically. The other packages will be marked as unavailable on
the first run, and can be marked foreign again easily with
unavailable-foreign.

If custom->foreign, foreign packages will still be marked as foreign,
yes, but frozen packages will be marked as installed and will be
ready for an upgrade, which is potentially dangerous. Specially in
the case of packages like aaa_elflibs.