Make the ".tgz" suffix part of the package data.
The ".tgz" suffix is now part of a SlackwarePackage object, in its
__suffix member. This commit adds the neccesary constants and upgrades
the constructors to pass it as part of the package data, and the
fullname() method has been upgraded to include it at the end. This
change makes it possible to call the fullname() method from a local
package and not get an spurious ".tgz" suffix at the end, while
packages read from FILELIST.TXT get it.

This is mostly important for the list-alternatives command output,
which had a small hack to avoid printing the ".tgz" suffix in local
packages. It is now safe to print the different versions with a
simple call to fullname().

This change, however, introduces a new member and a new method to
the SlackwarePackage objects. It's needed for fullname(), so I also
had to add a mechanism to force a rebuild in the local and remote
lists cache, because the cached SlackwarePackage objects wouldn't
have the __suffix member, resulting in errors. I have included the
new force-rebuild mechanism as part of the detection of a new
slackroll version from the "self" file. I have just tested it and
the package caches are rebuilt when increasing the version number,
so it shouldn't fail for other users.