Fix security problem in download_verify()
By downloading the packages directly to the packages cache, it was
possible to get evil packages in it. This could happen if slackroll
was interrupted (not by pressing Ctrl+C) after downloading the
package and before verifying the signature. This could happen if
the computer powered off suddenly or if someone killed the process
using signals, for example.

The fix introduces a new function called get_temp_dir() that is
used in two ways. First, the download() function now uses the
temporary directory in addition to the temporary suffix as the
temporary name for downloaded files. Second, the download_verify()
function downloads both the package and the signature to the
temporary directory and only after it has been verified it moves
the package file to the package cache. It is still possible in
theory to get an evil package into the package cache by overwriting
the temporary file in the temporary directory after the signature
has been verified and before the package is moved to the package
cache. This attack, while possible, is very hard to coordinate, and
I still don't know how to avoid it, if that's possible.