Flush stdout in choose_option() after prompting the user
This enables saving a list of URLs using the "tee" command
like:

	slackroll urls gcc | tee urls.txt

Previously, without flusing stdout, this command would not
output the questions to stdout before asking you for input.
With the fix it does, so it's usable this way. You can use
the command above and you'll see the questions, and tee
will save the output to urls.txt for you to edit.