Make glibc-solibs, sed and pkgtools the prioritized packages
Previously, the prioritized packages were the ones belonging to glibc.
With this change, some glibc packages are removed leaving only the
important one, glibc-solibs, and adds sed and pkgtools as the upgrading
documents always mention with new major releases. The concept of glibc
indicator also changes and disappears. Now, instead of checking if the
glibc indicator appears in a group of package names, the package names
are checked to see if they appear in the prioritized list.