Definitive version of the -stable package discarding algorithm
I could not get official word from Patrick Volkerding over this issue. I tried
to get expert advice or opinion about the best way to do this, from the people
at ##slackware on freenode but, again, I got nothing except Alien Bob trying to
be a rude smart ass for some reason. I suppose he was tired after I explained
the issue.

The problem is that, until not many commits ago, when I detected a package had a
version in the /patches directory, I discarded its package versions from the
main tree, in /slackware, and only from there. This doesn't work. Packages from
/extra get their updates, normally, in /patches. So not long ago, when a package
from /extra got an upgrade in /patches, the program didn't put it in the
outdated state, because it was considering the version from /extra still valid.
Aware of this problem, some commits ago I decided to discard both versions, from
/slackware and /extra, but this conflicts with an existing situation we have now
in Slackware 11.0.

In Slackware 11.0, the PHP package has a version in the main tree, another one
in extra and another one in patches. I suppose it was a bad choice from Patrick
Volkerding to call all those packages the same. The version from the main tree
is PHP 4. This version has its updates in /patches. The version from the extra
tree is PHP 5. This version has its updates in place. If he had put all the
updates in /patches, I could have kept this second version of the discarding
algorithm. But he didn't, as he explains in the Slackware 11.0 change log,
because "upgrade tools" would then wrongly take this version from /patches and
upgrade PHP 4 with PHP 5 without the user wanting it. So he decided to ship the
upgrades to PHP 5 in place. If, according to the version prior to this commit, I
remove the /extra and main versions, only the patches version would remain, and
the user would have lost the ability to run PHP 5, which is perfectly valid.
Like I said previously, this wouldn't have happened if Patrick Volkerding had
called the PHP 5 package "php5" instead of "php". Granted, this would have
created a minor problem, which is that in theory you could have both "php" and
"php5" installed at the same time and they weren't intended to coexist. This is
a minor issue, but he does with Slackware whatever he pleases, so if he pleases
to do something wrong and not well thought out, it's his right (and our
problem). I suppose he viewed the problem from the perspective of a "full
installation" and prefered to call them all the same so he didn't have to modify
anything and "upgradepkg" behaved at least half right.

Hence, this third iteration of the discarding algorithm, complying with all the
real world cases I could find. If a package has a version in the main tree and
in the extra tree, only the version from the main tree is discarded when patches
are present. This copes well with the "php" case. Otherwise, any version from
the main and extra trees are discarded. This copes well with the rest of patches
to packages from the main and extra trees, doesn't matter which.

This situation made me think a bit more about the problem, because we still have
the floating case of updates to /pasture and /testing. This is where Alien Bob
went rude after not giving any specific answer. The discarding algorithm I'm
commiting now will not work, as simple as that, if packages in /testing or
/pasture get their upgrades in /patches. They would have the same problem I
originally had with the packages from /extra not being discarded. All evidences
point to /testing receiving the upgrades in place. I could find some examples of
this, so I'm going to consider them the normal case, not an exceptional case
like the in place "php" upgrades in /extra. I could not remember upgrades to
/pasture packages. However, I don't use -stable so it's normal that I don't
remember /pasture upgrades going to /patches, even if they would have taken
place. For the moment, and I hope I'm not wrong, I'm going to suppose they would
be in-place, like /testing. If that's the case, everyone happy. The algorithm
I'm committing is the definitive and right one. If it's not, I hope some good
and patient user will mail me about the problem should we face it in the future.
Nobody was able to tell me if upgrades to /pasture were in-place. I don't know
if it's an absurd question or they simply didn't know and didn't want to
explicitly say they didn't know. If it's an absurd question, something like
"don't worry; packages in /pasture are never upgraded" would have been enough.

Done. I wanted to put this properly in words.