Always use slackware.osuosl.org to retrieve package signatures
After reading an article on slashdot.org some days ago about the security of
some popular package managers, I realized slackroll was susceptible to the
following malicious mirror scenario, among others:

1. User has foo-1.0 installed.
2. It's discovered that foo-1.0 has a security problem.
3. Upstream author fixes the problem and releases foo-1.1.
4. PV builds foo-1.1 and releases a signed security fix.
5. Evil mirror renames foo-1.0 and its signature into foo-1.1.
6. slackroll downloads the vulnerable package and its signature.
7. The signature appears valid.
8. User continues to run foo-1.0 thinking he has upgraded.
9. Evil mirror knows the IP address of a vulnerable user.

In theory, the people behind the evil mirror could attempt to attack the user if
an exploit is published and if the user doesn't change his IP address. They
could also pass a previous version of foo instead of using foo-1.0. Maybe
foo-0.5 had a known and exploitable security issue and they offer you that one
instead.

Despite any countermeasure we put in place, an evil mirror can operate for long
and obtain IP addresses of its users. Once an exploitable bug is found in a
package, the evil mirror could attack its users before they get the security
upgrade from them or while they're getting it. In this second scenario they
don't need to prevent you from getting foo-1.1, but just delay your upgrade long
enough.

We can't do anything about the second scenario, but we could do many things
about the first one. For example, some other package managers check both the MD5
of the package and its signature. The MD5 sums are available at the
CHECKSUMS.md5 files that populate the Slackware tree and they are signed too in
their CHECKSUMS.md5.asc files. Other package managers check the MD5 sums alone.
These two cases are less vulnerable to the type of attack we are describing
here, because at least the MD5 sums are linked to the file name. The GnuPG
signatures are not, which is the source of the problem.

While I could take some more severe measures for slackroll, I made a very
non-intrusive change (2 lines of code added, 2 changed). I'm forcing slackroll
to get the package signatures (not the packages themselves) from
slackware.osuosl.org, one of the primary FTP sites. Now the attackers would not
only need to set up an evil mirror, which is almost trivial if they are
committed to do it. They would also have to compromise slackware.osuosl.org
(where the rest of the mirrors pull from), which is much more difficult and it
would also be a pretty chaotic situation.