Check against files without .new suffix in .new files removal
The program suffered a bug that made it erroneously prompt for input on
.new files removal about files that were still present in installed
packages, due to the program failing to detect that they were still
present. An example:

If a new package had /some/dir/file.new, that was later moved to
/some/dir/file, and an old package had the same file, the program would
prompt you to keep or remove /some/dir/file, as it wouldn't realize that
it's present in a current package, only with the .new suffix.