Detect error exit codes when calling gpg2? --version
As reported by Vojtech Kupca, calling gpg2 or gpg may result in an error
with the programs being installed, that does not result in an IOError or
OSError exception in Python. One of these cases is when the program is
installed but its shared library dependencies are not met. In that case
the program is found, but, even if using --version should always work,
it does not. You get a nonzero exit code.