Fix wrong generator usage when detecting GnuPG
This commit fixes a long standing "bug" in slackroll. This is not a bug
per-se, because the code works, but it's simply not working the way it
was intended.

When detecting which version of GnuPG is installed, slackroll tries to
run gpg2 --version and gpg --version, and uses whichever succeeds first.
To avoid repeating this test everytime, the name of the program to use
was "yielded", and then slackroll used something like

yield_gnupg_exec_name().next()

everytime it wanted to run GnuPG. The problem is that when you create a
function that yields something, instead of returning a value, you get
some sort of iterator object. From it, you must call its "next()" method
to get the next value. If you do it the way described above, you would
be creating a new iterator object everytime. This implies running the
check everytime instead of just one time, which is what the code
attempted to accomplish. Naturally, performing the check each time does
work, but the code is not doing what it should.