Search for .new files in /etc too after installation
Previously, slackroll used the package file lists in /var/log/packages
to get a list of all .new files that had been added by installed
packages. However, as of the time of this commit, there are packages
that create .new files from the post-install script. Notably, php does
this:

cp -a etc/httpd/php.ini-production etc/httpd/php.ini.new
config etc/httpd/php.ini.new

As the install scripts are shell scripts, it's not practical to catch
these cases by parsing and understanding the script.

Other costly option is to try to catch, maybe using LD_PRELOAD, all .new
files that are created with installpkg, but this is a bit out of the
scope of slackroll, as currently it's a Python script and no binary
files are distributed with it. This would need C code and compiling bits
of the source in the slackbuild.

So we're left with a terribly easy option that works 99% of the time and
is what slackpkg does: simply find .new files in /etc after package
installation. slackroll, from now on, does that too *in addition* to the
.new search it was already doing. So we catch everything outside /etc
that's listed in the package contents, and also every .new file in /etc
that may be listed or not in the package contents.