Fix update-manifest operation
Previously, the update-manifest operation relied on
/var/slackroll/FILELIST.TXT being present when it was going to be run.
From that file, the program extracted the list of MANIFEST.bz2 files it
had to download and process.

Since the "update" operation now processes FILELIST.TXT just after
downloading it and then deletes it, the file is not available at
"update-manifest". The solution is to extract the list of MANIFEST.bz2
files when processing FILELIST.TXT during the "update" operation.

This list is then stored together with the mirror URL on disk in a file
called manifestlist.db that can be loaded with try_load. The data
structure has this format:

    [ (mirror_url, file_path), (mirror_url2, file_path2), ...]

The "update-manifest" operation now needs to load the list of
MANIFEST.bz2 files from manifestlist.db and proceed to download and
process them as it did before.