Start porting to support python3 - Stage1 fixes - `print`.
Following the guide on http://python-future.org/automatic_conversion.html

These are the safe fixes.

```
futurize --stage1 -w slackroll -f libfuturize.fixes.fix_print_with_import
```

I did not include the `from __future__ import print_function` since that
breaks Python 2.5 compatibility.

Now we could argue that we don't need to support Python 2.5 and anyone
who wanted to install slackroll on slackware < 14.0 can just use the
last supported version. That is true, but I don't need to break
compatibility here since every single instance of `print` in slackroll
is a single string, so we can wrap it with brackets and it behaves
exactly the same as before.

We might need to break compatibility later, but I've checked the current
script and it works on every slackware since 12.0 (after updating the
shebang back to `python` instead of `python2`.