Start porting to support python3 - Stage1 fixes - `reduce`.
Following the guide on http://python-future.org/automatic_conversion.html

These are the safe fixes.

```
futurize --stage1 -w slackroll -f lib2to3.fixes.fix_reduce
```

This simply imports the reduce function from functools, which is the
same function as reduce(). It is made available in this module to allow
writing code more forward-compatible with Python 3.

This support was added in python 2.6, so again, to keep compatibility
with Python 2.5, I only do it when we are not using Python 2.