Start porting to support python3 - Stage1 fixes - `next`.
Following the guide on http://python-future.org/automatic_conversion.html

These are the safe fixes.

```
futurize --stage1 -w slackroll -f libfuturize.fixes.fix_next_call
```

This changes calls to `.next()` on an iterator to use the `next`
builtin. Unfortunately again, this was added in Python 2.6. Again to
keep compatibility, I've fallen back and defined a `next` function to
handle this for us.

Additionally, this need to have the caching of the gpg binary lookup
adjusted. We are trying to avoid computing which to use if we are
checking multiple signatures, see 24afea0 for the initial details when
this was introduced.

I've made a global for this so we can continue to cache the value.