Setup poetry, add requirements for pytest.
I've wrapped the "main" method with `if __name__ == '__main__':` to
allow importing this file for testing.

This needs some additional refactoring of the parts in main to make it
simpler to test, but this will get us started and allow us to write some
tests to begin the python3 porting work.