Wrap "main" code with `if __name__ == '__main__':` to allow testing.
This is cherry-picked from 42c9909 so we can have a stable set of tests
for the python2 branch before we fork off.

To allow setting up some tests for slackroll, we need to be able to
import things from this file without running it.

This needs some additional refactoring of the parts in main to make it
simpler to test, but this will get us started and allow us to write some
tests to begin the python3 porting work.