Setup pre-commit for checks.
Run black and isort for consistent formatting and add checks for them to
github action.

It also verifies our types using mypy, pinned to the last version which
supports python2.

I did not yet blacken `slackroll` since that will make a huge diff and
there is probably not a lot of value in doing that right now. This will
however keep the tests well formatted going forward.