Mop up final stage2 fixes. (#45)
Following the guide on https://python-future.org/automatic_conversion.html#stage-2-py3-style-code-with-wrappers-for-py2

I've now applied all the stage2 fixes.

We'll iterate on tests and python3 now and start pulling in changes from
@spillner.

futurize is installed by default on 15.0. We might want to support older
slackware versions, but first we will get everything working on 15.0
with python2 and python3 and good test coverage. We can still branch out
to testing with current and potentially older releases. (which might
mean dropping the futurize dependency again).