Switch on-disk format of persistent.db. (#47)
Internally, `shelve` uses `anydbm.open` to open the file:

https://github.com/python/cpython/blob/v2.7.18/Lib/anydbm.py#L57-L85

On Python2, this defaults to `dbhash`:

https://github.com/python/cpython/blob/v2.7.18/Lib/anydbm.py#L39

dbhash is deprecated and has been removed in Python3.

https://github.com/python/cpython/blob/v2.7.18/Lib/dbhash.py

To handle this problem we have a couple of options. We could make
slackroll depend on an external package, berkeleydb, and we could then
open the file.

https://pypi.org/project/berkeleydb/

Historically though, slackroll has been a single python script with no
external dependencies and if possible, I think I would like to keep it
that way.

So another option, is to change the on-disk format on the fly so that
when we do eventually ship a new version of slackroll which uses
python3, everyone's persistent.db file will have a format that can be 
read by that version and everything just works.

This is the option I am implementing in this change. This will switch
the file to use gnu dbm as the format instead.

Thanks to @spillner for all the investigation in #23, so I could
understand what was changing.