Reuse existing get_temp_dir + try_to_rename for upgrading db.
In 2356f7f, I added code to switch the on-disc format used by
peristence.db from dbhash to gnu dbm. I used mkdtemp to create a temp dir
and created a new file in gnu dbm format and then renamed the file with
`os.rename` to overwrite the existing file.

I missed that slackroll already has a helper `get_temp_dir` for such
file operations, but I did notice that is uses `os.rename` elsewhere.

To avoid duplication, I will switch to the existing helper (which uses
/var/slackroll/tmp as the tmp dir if the user does not have a TMPDIR env
var set.

There is still an issue if that env var is set and the dir is on a
different filesystem. I will handle this in a second commit.

Closes #52