Setup poetry, add requirements for pytest.
This is cherry-picked from 42c9909 so we can have a stable set of tests
for the python2 branch before we fork off.

This is using poetry 1.1.15, the last version that allowed poetry to run
with python 2.7.

I will document details later of how to set it up so we can test this
python2 code, but for now, this will allow us to setup some tests before
we start with the python 3 porting work.