Avoid reusing variable names to allow mypy type-checking.
I think this should actually not be a problem with more modern tools -
pyright does not seem to actually mind this, but this version of mypy
flags things where you assign a variable with some type and then reuse
the variable with something of a completely different type later.

This simply makes the names unique so we can enable type-checking to
fail as a pre-commit check.