Alias anydbm to dbm on python2 for forward compatibility.
We've previously done the work to have python2 upgrade the on-disc
format from bsddb to gdbm so that we can cutover to python3 in the
future.

This change simply guards the python2 code and aliases dbm to point
to anydbm. There should be no change to python2, but this will allow
work to begin on python3 db testing.